split up modules and added tests
