+++ tests/helpers/data/words.csv
@@ -1,0 +1,29 @@
+"@testaccount",
+"@testaccounthypen_1",
+"@testaccountnum1",
+"a",
+"a.nd",
+"bad","bad"
+"characters","characters"
+"go",
+"great","great"
+"here",
+"http://te.gov",
+"https://test.co",
+"in",
+"nuke","nuke"
+"numb3r$","numbr"
+"other",
+"remove","remove"
+"so_me","so_me"
+"some",
+"test","test"
+"test+2@gmail.co",
+"test-2@email.com",
+"test@email.com",
+"test_12@ema-il.org",
+"ttest++sdfsdflkj34+sdfa_sdf@gmail.co",
+"tweet","tweet"
+"w1th","wth"
+"words","words"
+"words1","words"
